First commited as after_11094_rev1

+++ perf/org.eclipse.linuxtools.perf/src/org/eclipse/linuxtools/internal/perf/model/PMSymbol.java
@@ -27,7 +27,7 @@ public class PMSymbol extends TreeParent {
 		return tmpName;
 	}
 
-	public PMSymbol(String symbolName, float pc, int samples) {
+	public PMSymbol(String symbolName, float pc, double samples) {
 		super(symbolName, pc, samples);
 	}
